+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bug fixes
   - `alt+shift+c` not working when pressed before dragging
+  - incremental pasting issue with address property
 
 - editor
   - add `mod+d` and `mod+shift+d` for duplicating widgets